refactor: moves buildpatterns and resolveMatrixConfig into a single util.ts files
