+++ src/configs/js.ts
@@ -8,7 +8,7 @@ import tsParser from '@typescript-eslint/parser';
 import { FlatCompat } from '@eslint/eslintrc';
 import { fixupPluginRules } from '@eslint/compat';
 import matrixaiPlugin from '../index.js';
-import { resolveMatrixConfig } from '../utils/resolveMatrixConfig.js';
+import { resolveMatrixConfig } from '../utils/utils.js';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 const __filename = fileURLToPath(import.meta.url);
